requirements.txt
# This file describes the distributions required to run this project
# in all envorinments.
# See http://pip-installer.org/requirement-format.html for details.

Django==1.8.3
djangorestframework==3.1.2
